+++ src/Support/Enums/HttpResponseCodesEnum.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Orizaba\LaravelComponents\Support\Enums;
+
+enum HttpResponseCodesEnum: int
+{
+    case CONTINUE = 100;
+    case OK = 200;
+    case CREATED = 201;
+    case ACCEPTED = 202;
+    case NO_CONTENT = 204;
+    case RESET_CONTENT = 205;
+    case PARTIAL_CONTENT = 206;
+    case MULTIPLE_CHOICES = 300;
+    case MOVED_PERMANENTLY = 301;
+    case FOUND = 302;
+    case SEE_OTHER = 303;
+    case NOT_MODIFIED = 304;
+    case BAD_REQUEST = 400;
+    case UNAUTHORIZED = 401;
+    case PAYMENT_REQUIRED = 402;
+    case FORBIDDEN = 403;
+    case NOT_FOUND = 404;
+    case METHOD_NOT_ALLOWED = 405;
+    case NOT_ACCEPTABLE = 406;
+    case REQUEST_TIMEOUT = 408;
+    case CONFLICT = 409;
+    case GONE = 410;
+    case INTERNAL_SERVER_ERROR = 500;
+    case NOT_IMPLEMENTED = 501;
+}